Update README, add sandbox mode, improve error handling
The README now includes more helpful sample responses and request, as
well as a table of contents

Sandboxing is enabled. Set the Sandbox header in your request to "true",
and the API will swap in Assembly's test secret key; the key is reset on
each request.

Error handling also got a bit more informative for Stripe errors.